implement auth using policy

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,10 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+    public function __construct() {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     //@return \Illuminate\Http\Response
     public function index() {
         //  \DB::enableQueryLog();
@@ -40,16 +44,20 @@ class QuestionsController extends Controller
     }
 
     public function edit(Question $question) {
+        $this->authorize("update", $question);
+
         return view("questions.edit", compact('question'));
     }
 
     public function update(AskQuestionRequest $request, Question $question) {
+        $this->authorize("update", $question);
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', 'Your question has been updated.');
     }
 
     public function destroy(Question $question) {
+        $this->authorize("delete", $question);
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your question has been deletec.');
